Completed test refactor for gpii.express module per feedback from Antranig.
