+++ tests/js/index.js
@@ -1,3 +1,5 @@
+"use strict";
+require("./test-caseholder");
 require("./test-middleware");
 require("./test-router");
 require("./test-router-cookiesetter");
